pyproject.toml
[project]
name = "ludamus"
version = "0.1.0"
description = "Event management website"
authors = [
    { name = "Radosław Ganczarek", email = "radoslaw@ganczarek.in" },
    { name = "hasparus", email = "hasparus@gmail.com" },
    { name = "lprasol", email = "lprasol@gmail.com" },
]
license = { text = "BSD-3-Clause license" }
readme = "README.md"
requires-python = ">=3.14.0,<3.15"
dynamic = ['dependencies']

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
authlib = "~1.6.0"
django = "~6.0.0"
django-bootstrap5 = "~26.1.0"
gunicorn = "^25.0.1"
django-environ = "^0.12.0"
psycopg = "^3.2.9"
pydantic = "^2.11.7"
django-extensions = "^4.1"
requests = "^2.32.5"
heroicons = { version = "^2.9", extras = ["django"] }
django-tailwind = { version = "^4.0.0", extras = ["reload"] }

[tool.poetry.group.dev.dependencies]
black = "^26.1.0"
codespell = "^2.4.1"
deptry = "^0.24.0"
django-debug-toolbar = "^6.0.0"
django-stubs = "^5.2.0"
djlint = "^1.36.4"
factory-boy = "^3.3.3"
faker = "^40.1.0"
freezegun = "^1.5.2"
ipython = "^9.9.0"
mypy = "^1.16.0"
pip-audit = "^2.10.0"
pre-commit = "^4.5.1"
pylint = "^4.0.3"
pytest = "^9.0.1"
pytest-cov = "^7.0.0"
pytest-django = "^4.11.1"
pytest-factoryboy = "^2.8.1"
pytest-freezeblaster = "^2.0.4"
responses = "^0.25.8"
ruff = "^0.15.0"
types-authlib = "^1.5.0.20250516"
types-requests = "^2.32.4.20250809"

# BLACK

[tool.black]
line-length = 88
preview = true
skip-magic-trailing-comma = true
target-version = ['py314']

# CODESPELL

[tool.codespell]
skip = 'node_modules,tests/e2e,*.po,*.lock,package-lock.json,vendor,src/ludamus/gates/web/django/theme/static_src/package-lock.json'

# COVERAGE

[tool.coverage.run]
branch = true
source = ["src", "tests"]
omit = [
    "src/ludamus/edges/*",
    "src/ludamus/adapters/db/django/migrations/*",
    "tests/template_checks.py",
]

[tool.coverage.report]
show_missing = true
skip_covered = true
exclude_also = ["if TYPE_CHECKING:", "# pragma: no cover"]
precision = 2

# DJLINT

[tool.djlint]
ignore = "H021,H031,H037,D018"

# DEPTRY

[tool.deptry]
known_first_party = ["ludamus"]

[tool.deptry.per_rule_ignores]
DEP002 = [
    "django-bootstrap5",
    "django-extensions",
    "django-tailwind",
    "gunicorn",
    "heroicons",
    "psycopg",
]
DEP004 = ["debug_toolbar"]

# MYPY

[tool.mypy]
check_untyped_defs = true
color_output = true
disallow_any_decorated = true
disallow_any_explicit = false                 # TODO: REMOVE
disallow_any_expr = false
disallow_any_generics = true
disallow_any_unimported = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
exclude = ["migrations/.*.py", "tests/.*.py"]
python_version = "3.14"
plugins = ["mypy_django_plugin.main"]
warn_incomplete_stub = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.django-stubs]
django_settings_module = "ludamus.edges.settings"

[[tool.mypy.overrides]]
module = ["debug_toolbar.toolbar"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["environ"]
ignore_missing_imports = true

# PYTEST

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "ludamus.edges.settings"

# RUFF

[tool.ruff]
fix = false
line-length = 88
output-format = 'concise'
preview = true

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    # Unused
    "COM",
    "CPY",
    "D1",
    "FIX",
    "TD003",
    # Left to fix
    "PLR0911", # Too many return statements
    "PLR0912", # Too many branches
    "PLR0915", # Too many statements
    "C901",    # Too complex
    "PLR0914", # Too many local variables
    "PLR1702", # Too many nested blocks
    "EM101",   # Exception must not use a string literal, assign to variable first
    "TRY003",  # Avoid specifying long messages outside the exception class
]

[tool.ruff.lint.per-file-ignores]
"*/migrations/**.py" = [
    "ANN",    # flake8-annotations
    "ARG001", # Unused arguments
    "E501",   # Line too long
    "RUF012", # ClassVar
]
"tests/**.py" = [
    "ANN",     # flake8-annotations
    "INP001",  # implicit namespace package
    "PLR0913", # too-many-arguments
    "PLR0917", # too-many-positional-arguments
    "S",       # bandit
    "PLR0904", # Too many public methods
    "PLR6301", # Method could be a function
]

[tool.ruff.lint.flake8-pytest-style]
parametrize-values-type = "tuple"

[tool.ruff.lint.flake8-type-checking]
runtime-evaluated-base-classes = ["pydantic.BaseModel"]

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.lint.pydocstyle]
convention = "pep257"


# IMPORT LINTER

[tool.importlinter]
include_external_packages = "True"
root_package = "ludamus"

# Explanation <source_modules> can't import <forbidden_modules>

[[tool.importlinter.contracts]]
name = "gates"
type = "forbidden"
source_modules = ["ludamus.gates"]
forbidden_modules = [
    "ludamus.links",
    "ludamus.inits",
    "ludamus.specs",
    "ludamus.edges",
]


[[tool.importlinter.contracts]]
name = "links"
type = "forbidden"
source_modules = ["ludamus.links"]
forbidden_modules = [
    "ludamus.gates",
    "ludamus.inits",
    "ludamus.specs",
    "ludamus.edges",
]


[[tool.importlinter.contracts]]
name = "inits"
type = "forbidden"
source_modules = ["ludamus.inits"]
forbidden_modules = ["ludamus.edges"]


[[tool.importlinter.contracts]]
name = "mills"
type = "forbidden"
source_modules = ["ludamus.mills"]
forbidden_modules = [
    "ludamus.gates",
    "ludamus.links",
    "ludamus.inits",
    "ludamus.specs",
    "ludamus.edges",
    "django",
]


[[tool.importlinter.contracts]]
name = "pacts"
type = "forbidden"
source_modules = ["ludamus.pacts"]
forbidden_modules = [
    "ludamus.gates",
    "ludamus.links",
    "ludamus.inits",
    "ludamus.mills",
    "ludamus.specs",
    "ludamus.edges",
    "django",
]


[[tool.importlinter.contracts]]
name = "specs"
type = "forbidden"
source_modules = ["ludamus.specs"]
forbidden_modules = [
    "ludamus.gates",
    "ludamus.links",
    "ludamus.inits",
    "ludamus.mills",
    "ludamus.edges",
    "django",
]


[[tool.importlinter.contracts]]
name = "edges"
type = "forbidden"
source_modules = ["ludamus.edges"]
forbidden_modules = [
    "ludamus.gates",
    "ludamus.links",
    "ludamus.inits",
    "ludamus.mills",
    "ludamus.pacts",
    "ludamus.specs",
]
